Added drag and drop UI for treeView.
Drag and drop events are now handled and the new code can determine the
source (what is dragged) and the destination (where it is dragged). The
actual movement of tasks is not enabled yet, but can be implemented in
the future by simple adding a function that does the move.