Added task tree locking.
When a task tree is opened, a lock file is created in the root directory
of the task tree. If the file already exists, the tree is not opened
unless the user decides to ignore the lock file. The lock file is
deleted when the virtual machine or the TaskStore is closed.